authentication flow, github oauth, github code web, authentication route, jwt token, saving token in cookies, displaying user profile

+++ server/src/routes/memories.ts
@@ -4,8 +4,15 @@ import { z } from "zod";
 
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify();
+  })
+
+  app.get("/memories", async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: "asc",
       },
@@ -20,7 +27,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     });
   })
 
-  app.get("/memories/:id", async (request) => {
+  app.get("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -31,7 +38,11 @@ export async function memoriesRoutes(app: FastifyInstance) {
       where: {
         id,
       },
-    })
+    });
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    };
 
     return memory;
   })
@@ -50,14 +61,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '',
+        userId: request.user.sub,
       },
     })
 
     return memory;
   })
 
-  app.put("/memories/:id", async (request) => {
+  app.put("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -72,7 +83,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    };
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -86,12 +107,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory;
   })
 
-  app.delete("/memories/:id", async (request) => {
+  app.delete("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(request.params);
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    };
 
     await prisma.memory.delete({
       where: {
